Added password hashing ; Added jwt tokenization, api endpoints now depend on loged in user's token

+++ src/backend/services/todo.py
@@ -5,17 +5,21 @@ class ToDo:
 
     # GET Operations
 
-    def todo_list(db:Session):
+    def todo_list_all(db:Session):
         return db.query(Todos).all()
 
     def todo_by_id(db: Session, todo_id:int):
         return db.query(Todos).filter(Todos.id==todo_id).first()
 
+    def user_todo_list(db: Session, owner_id: int):
+        return db.query(Todos).filter(Todos.owner==owner_id).all()
+
 
     # POST Operations
 
-    def create_todo(db: Session, body):
+    def create_todo(db: Session, body, owner_id: int):
         todo_model = Todos(**body.model_dump())
+        todo_model.owner= owner_id
 
         db.add(todo_model)
         db.commit()
@@ -24,11 +28,10 @@ class ToDo:
 
     # PUT Operations
 
-    def update_todo(db: Session, body, todo_id: int):
-        todo_model= db.query(Todos).filter(Todos.id==todo_id).first()
+    def update_todo(db: Session, body, todo_id: int, owner_id: int):
+        todo_model= db.query(Todos).filter((Todos.owner==owner_id) & (Todos.id==todo_id)).first()
         if todo_model is None:
             return False
-
         updates = body.model_dump(exclude_unset=True, exclude_none=True)
 
         for field, value in updates.items():
@@ -40,11 +43,11 @@ class ToDo:
 
     # DELETE Operations
 
-    def delete_todo(db: Session, todo_id: int):
-        todo_model=db.query(Todos).filter(Todos.id==todo_id).first()
+    def delete_todo(db: Session, todo_id: int, owner_id: int):
+        todo_model=db.query(Todos).filter((Todos.owner==owner_id) & (Todos.id==todo_id)).first()
 
         if todo_model is None:
             return False
-        db.query(Todos).filter(Todos.id==todo_id).delete()
+        db.query(Todos).filter((Todos.owner==owner_id) & (Todos.id==todo_id)).delete()
         db.commit()
         return True
